+++ COT/tests/test_doctests.py
@@ -34,8 +34,5 @@ def load_tests(*_):
     suite = TestSuite()
     suite.addTests(DocTestSuite('COT.data_validation'))
     suite.addTests(DocTestSuite('COT.file_reference'))
-    suite.addTests(DocTestSuite('COT.logging_'))
-    suite.addTests(DocTestSuite('COT.platforms'))
     suite.addTests(DocTestSuite('COT.utilities'))
-    suite.addTests(DocTestSuite('COT.vm_description'))
     return suite