+++ COT/ui/ui.py
@@ -27,11 +27,6 @@
 import logging
 import sys
 
-from verboselogs import VerboseLogger
-
-# VerboseLogger adds a log level 'verbose' between 'info' and 'debug'.
-# This lets us be a bit more fine-grained in our logging verbosity.
-logging.setLoggerClass(VerboseLogger)
 logger = logging.getLogger(__name__)